+++ src/main/java/application/model/records/services/IRecordsRepository.java
@@ -1,7 +1,7 @@
 package application.model.records.services;
 
 import application.model.records.Record;
-
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 import java.util.Map;
@@ -9,5 +9,5 @@ import java.util.Map;
 public interface IRecordsRepository {
     List<Record> getProducts(Timestamp timestamp);
     Map<String,List<Float>> getLatestPrices(List<String> productNames);
-    void updatePrices(List<Map.Entry<String, Float>> data, String invoiceDate);
+    void updatePrices(List<Map.Entry<String, BigDecimal>> data, String invoiceDate);
 }
